factor out d2graph/layout.go
